// SERVER/src/filters/enums/error-code.ts
export enum ErrorCode {
  E400001 = 'E400001',

  E400002 = 'E400002',

  E400003 = 'E400003',

  E400004 = 'E400004',

  E400005 = 'E400005',

  E400006 = 'E400006',

  E400007 = 'E400007',

  E400008 = 'E400008',

  E500001 = 'E500001',

  E500002 = 'E500002',

  E500003 = 'E500003',

  E500004 = 'E500004',

  E500005 = 'E500005',

  E500006 = 'E500006',

  E500007 = 'E500007',

  E405002 = 'E405002',

  E405003 = 'E405003',
}